move resource unit test
